10, 118, 12, 496, 606, 63

+++ 10/k3.c
@@ -4,16 +4,16 @@ int main() {
     freopen("input.txt", "r", stdin);
     //freopen("output.txt", "w", stdout);
  
-	long long r, a, b, c, d;
-    scanf("%lld %lld %lld %lld", &a, &b, &c, &d);
+	double r, a, b, c, d;
+    scanf("%lf %lf %lf %lf", &a, &b, &c, &d);
  
-    long long x, from = -100, to = 100;
+    long x, from = -100, to = 100;
      
     for (x = from; x <= to; x++) {
         r = a*x*x*x + b*x*x + c*x + d;
-        //printf("%ld : %d*x^3 + %d*x^2 + %d*x + %d = %lld\n", x, a, b, c, d, r);
+        //printf("%ld : %.0lf*x^3 + %.0lf*x^2 + %.0lf*x + %.0lf = %.0lf\n", x, a, b, c, d, r);
         if (r == 0)
-			printf("%lld ", x);
+			printf("%ld ", x);
     }
  
     return 0;
